Blaze fireball recharge speed config
Slight change to how this works internally, but the end result and the default functionality are unchanged. This option is presented to the user as a multiplier to how quickly Blazes' fireball attack recharges. Nox had been reducing the cooldown by 1 extra tick with each tick, doubling the speed at which the Blaze's Fireball recharges, so this config flag defaults to a multiplier of 2.

Does not actually multiply anything.