Oh what a tangled web we weave...
...when first we mess with NBT.

Use blursed sorcery to track which Cobwebs are placed by Spiders. A Spider can have a maximum of 5 Cobwebs placed in the world. A Cobweb persists for 20 seconds, but can stay in the world longer if there's a LivingEntity caught in it. If a Spider has 5 Cobwebs placed and attempts to place another, the Cobweb farthest from its target will be removed. All of a Spider's Cobwebs are removed from the world upon its death. Attempting to harvest a Spider-placed Cobweb will yield no drops to the player. Spider-placed Cobweb coordinates and ages are stored in the Spider's NBT when it's unloaded.