Merged Config Options (#25)
Many, many thanks to @SrGnis , @AeiouEnigma , and @tesseract-two !

* update (fix?) yarn mappings build

resolved a gradle build error I was having :)

* Refactor subclasses of MobEntityMixin to all override nox$initGoals instead of implementing their own injects, where applicable

* Modify comment on friendlyFire config flag to clarify that it applies only to mobs

I'm nitpicky tbh

* Change block breaking config flag to be configurable per-mob, with configurable block hardness cap

* Optimize Imports

* Fix Cave Spiders not inheriting Spider nox$onSuccessfulAttack changes

* Make Spider web placement configurable

* Configurable Blaze ignition radius

* Configurable whether Monsters are provoked by contact

* Skeleton swim config

* Blaze fireball recharge speed config

Slight change to how this works internally, but the end result and the default functionality are unchanged. This option is presented to the user as a multiplier to how quickly Blazes' fireball attack recharges. Nox had been reducing the cooldown by 1 extra tick with each tick, doubling the speed at which the Blaze's Fireball recharges, so this config flag defaults to a multiplier of 2.

Does not actually multiply anything.

* Move config check in Nox$CreeperBreachGoal::canStart to method head to avoid potentially unnecessary call CreeperEntity::getTarget

* Replace boolean buffAllGolems and buffAllMonsters config flags with multiplier variables for monster/golem health and follow range.

* Remove redundant blazesIgniteNearbyTargets config flag

* Allow configuring Zombie knockback resistance and pounce behavior. Also allow disabling Cave Spider slowness application.

* Undo mistakenly making Cave Spiders create Webs *and* apply Slowness. They now only apply Slowness, unless the new config flag to let them place webs is enabled

* Enderman and Creeper config options

* Slime-related configs

* change Mixin compatibilityLevel to JAVA_17

* configurable natural spawn weight for Guardians

* Oh what a tangled web we weave...

...when first we mess with NBT.

Use blursed sorcery to track which Cobwebs are placed by Spiders. A Spider can have a maximum of 5 Cobwebs placed in the world. A Cobweb persists for 20 seconds, but can stay in the world longer if there's a LivingEntity caught in it. If a Spider has 5 Cobwebs placed and attempts to place another, the Cobweb farthest from its target will be removed. All of a Spider's Cobwebs are removed from the world upon its death. Attempting to harvest a Spider-placed Cobweb will yield no drops to the player. Spider-placed Cobweb coordinates and ages are stored in the Spider's NBT when it's unloaded.

* Un-break several things :)

* Housekeeping

* Rework goal-related config flags such that the config check is done in the goal's canStart method

* tidiness >:(

* Spider fall damage immunity is configurable. Other minor tidying.

* Correct Cave Spider slowness documentation to reflect that they apply Slowness III, not Slowness II

* Bulk commit, read on

The biggest change in this commit is the addition of Polymer as an included library dependency. Polymer enables the creation of new server side blocks in a way that's entirely compatible with vanilla clients. The old Spider attack Cobweb placement code has been stripped out and replaced with a much simpler implementation based using Polymer. Spiders now place the new `nox:cobweb` block, which has a random chance to despawn every handful of ticks after it has existed for 10 seconds, provided that there's no LivingEntity caught in it. On average, they should disappear after about 20 seconds. There is no longer a limit to how many Cobwebs a Spider can have in the world at a time.

Other smaller changes include:

Config options for Nether spawns
Config options for non-Nether spawns
Witch-related options
Slime-related options
Option to disable Magma Cube lava placement
Option for Magma Cubes to place non-source lava instead of source lava
Enderman teleportation on melee hit config
Spider fall damage immunity config

* Wither Skeleton configs

* Vindicator/Evoker/Ghast config settings, other cleanup, and config toggle for magma cube igniting target on hit

* Config options for Ender Dragon, Endermite, Guardian water placement, Husk hunger application, Iron/Snow Golems, Phantoms, Shulker blindness application, Silverfish, Stray slowness application, and Wither.

* Perform null checks on EntityAttributeInstance modifiers

Also add EntityType checks to SlimeEntityMixin and SpiderEntityMixin to try to avoid poison clouds and web placement from mod-added slimes and from mod-added spiders which are smaller than a Cave Spider.

* Set Wither to full health in tail of WitherEntity::onSummoned

Closes #7

* Update Polymer

Remove placeholder block that's no longer needed, since Patbox fixed the relevant bug :)

* Improve configurability, more Snow Golem options, witherSuffocate rule, fix End Crystal immunity, add Always Mineable and Can't Mine tags. Copied lots from AeiouEnigma!

* more fixes

* some compilati0n fixes

* fix: last compilation errors

* refactor: some config refactor

* feat: add enderDragonBaseHealthMultiplier config option

* half config fix

* feat: all configs fixed

* fix: typo in lang

* build: bump version to match the version in the releases

* fix: config descriptions moved to tooltips

* fix: update license headers

* build: add mc version

* fix: creepersAttackShields not working

* build: version bump

---------

Co-authored-by: Aeiou <3160746+AeiouEnigma@users.noreply.github.com>
Co-authored-by: clearKeep <118913639+tesseract-two@users.noreply.github.com>
Co-authored-by: Genis Martinez <genis.martinez@biotechvana.com>